[Bugfix] Could not add images to popup content. [MER-1615]
