+++ assets/src/components/editing/elements/popup/PopupElement.scss
@@ -27,5 +27,16 @@
 .popup__modalContent {
   text-align: initial;
   min-width: 600px;
-  min-height: 485px;
+}
+
+.popup__content_editor {
+  color: authoring.$body-color;
+
+  margin-left: 30px;
+  border: 1px dashed;
+  padding: 10px;
+
+  h2 {
+    font-size: 1.2rem;
+  }
 }
